Use a dedicated navigation for training pages
In order to display a progress indicator for course on the navigation,
we do the following for each entry:

1. fetch the associated page from the content catalog
2. get the attribute `page-slug`
3. add a data attribute `data-slug` on the `<li>` element

The main issue is that fetching a page from the content catalog is relatively slow and the function will be called a lot.
For instance, if we have a simple course with 5 pages and each page has 5 entries in the navigation.
Then we will call `page-by-url` 5 * 5 = 25 times.